+++ packages/adminkit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gvray/adminkit
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @gvray/eskit@1.6.0
+
 ## 1.3.0
 
 ### Minor Changes