+++ packages/domkit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gvray/domkit
 
+## 1.4.0
+
+### Minor Changes
+
+- add new function
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @gvray/eskit@1.6.0
+
 ## 1.3.0
 
 ### Minor Changes